% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BED.R
\name{BED}
\alias{BED}
\title{Biological Entity Dictionary (BED)}
\description{
An interface for the neo4j database providing
mapping between different identifiers of biological entities.
\itemize{
\item \href{../doc/BED.html}{Vignette}
\item URL: \url{https://github.com/patzaw/BED}
\item BugReports: \url{https://github.com/patzaw/BED/issues}
}
}
\author{
Patrice Godard
}
