% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedImport.R
\name{bedImport}
\alias{bedImport}
\title{Feeding BED: Imports a data.frame in the BED graph database}
\usage{
bedImport(cql, toImport, periodicCommit = 10000, ...)
}
\arguments{
\item{cql}{the CQL query to be applied on each row of toImport}

\item{toImport}{the data.frame to be imported as "row".
Use "row.FIELD" in the cql query to refer to one FIELD of the toImport
data.frame}

\item{periodicCommit}{use periodic commit when loading the data
(default: 1000).}

\item{...}{additional parameters for \link{bedCall}}
}
\value{
the results of the query
}
\description{
Not exported to avoid unintended modifications of the DB.
}
\seealso{
\link{bedCall}, \link[neo2R:import_from_df]{neo2R::import_from_df}
}
