% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEIDList.R
\name{BEIDList}
\alias{BEIDList}
\title{Create a BEIDList}
\usage{
BEIDList(l, metadata, scope)
}
\arguments{
\item{l}{a named list of BEID vectors}

\item{metadata}{a data.frame with rownames or
a column "\strong{.lname}" all in names of l}

\item{scope}{a list with 3 character vectors of length one named "be",
"source" and "organism"}
}
\value{
A BEIDList object which is a list of BEID vectors with 2 additional
attributes:
\itemize{
\item \strong{metadata}: a tibble with metadata about list elements.
The "\strong{.lname}" column correspond to the names of the BEIDList.
\item \strong{scope}: the BEID scope ("be", "source" and "organism")
}
}
\description{
Create a BEIDList
}
