% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{BEIDs}
\alias{BEIDs}
\title{Get the BEIDs from an object}
\usage{
BEIDs(x, ...)
}
\value{
A tibble with at least 4 columns:
- value
- be
- source
- organism
- ...
}
\description{
Get the BEIDs from an object
}
