% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beIDsToAllScopes.R
\name{beIDsToAllScopes}
\alias{beIDsToAllScopes}
\title{Find all BEID and ProbeID corresponding to a BE}
\usage{
beIDsToAllScopes(
  beids,
  be,
  source,
  organism,
  entities = NULL,
  canonical_symbols = TRUE,
  entity_warning = TRUE
)
}
\arguments{
\item{beids}{a character vector of gene identifiers}

\item{be}{one BE. \strong{Guessed if not provided}}

\item{source}{the source of gene identifiers. \strong{Guessed if not provided}}

\item{organism}{the gene organism. \strong{Guessed if not provided}}

\item{entities}{a numeric vector of gene entity. If NULL (default),
beids, source and organism arguments are used to identify BEs.
Be carefull when using entities as these identifiers are not stable.}

\item{canonical_symbols}{return only canonical symbols (default: TRUE).}

\item{entity_warning}{by default (TRUE) a warning is shown when
the entities argument is used. Set this argument to FALSE to avoid this
warning.}
}
\value{
A data.frame with the following fields:
\itemize{
\item \strong{value}: the identifier
\item \strong{be}: the type of BE
\item \strong{organism}: the BE organism
\item \strong{source}: the source of the identifier
\item \strong{Gene_entity}: the gene entity input
\item \strong{GeneID} (optional): the gene ID input
\item \strong{Gene_source} (optional): the gene source input
\item \strong{Gene_organism} (optional): the gene organism input
}
}
\description{
Find all BEID and ProbeID corresponding to a BE
}
