% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheBedCall.R
\name{cacheBedCall}
\alias{cacheBedCall}
\title{Cached neo4j call}
\usage{
cacheBedCall(..., tn, recache = FALSE)
}
\arguments{
\item{...}{params for \link{bedCall}}

\item{tn}{the name of the cached table}

\item{recache}{boolean indicating if the CQL query should be run even if
the table is already in cache}
}
\value{
The results of the \link{bedCall}.
}
\description{
This function calls neo4j DB the first time a query is sent and puts
the result in the cache SQLite database. The next time the same query is
called, it loads the results directly from cache SQLite database.
}
\details{
Use only with "row" result returned by DB request.

Internal use.
}
\seealso{
\link{cacheBedResult}, \link{bedCall}
}
