% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheBedResult.R
\name{cacheBedResult}
\alias{cacheBedResult}
\title{Put a BED query result in cache}
\usage{
cacheBedResult(value, name)
}
\arguments{
\item{value}{the result to cache}

\item{name}{the name of the query}
}
\description{
Internal use
}
\seealso{
\link{cacheBedCall}, \link{loadBedResult}
}
