% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBedCache.R
\name{checkBedCache}
\alias{checkBedCache}
\title{Check BED cache}
\usage{
checkBedCache(newCon = FALSE)
}
\arguments{
\item{newCon}{if TRUE for the loading of the system information file}
}
\description{
This function checks information recorded into BED cache
and resets it if not relevant.
}
\details{
Internal use.
}
\seealso{
\link{clearBedCache}, \link{lsBedCache}
}
