% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBedConn.R
\name{checkBedConn}
\alias{checkBedConn}
\title{Check if there is a connection to a BED database}
\usage{
checkBedConn(verbose = FALSE)
}
\arguments{
\item{verbose}{if TRUE print information about the BED connection
(default: FALSE).}
}
\value{
\itemize{
\item TRUE if the connection can be established
\item Or FALSE if the connection cannot be established or the "System" node
does not exist or does not have "BED" as name or any version recorded.
}
}
\description{
Check if there is a connection to a BED database
}
\seealso{
\link{connectToBed}
}
