% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDubiousXRef.R
\name{cleanDubiousXRef}
\alias{cleanDubiousXRef}
\title{Identify and remove dubious cross-references}
\usage{
cleanDubiousXRef(d, strict = TRUE)
}
\arguments{
\item{d}{a cross-reference data.frame with 2 columns.}

\item{strict}{if TRUE (default), the function returns only unambiguous
mappings}
}
\value{
This function returns d without dubious cross-references.
Issues are reported in attr(d, "issues").
}
\description{
Not exported to avoid unintended modifications of the DB.
}
