% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearBedCache.R
\name{clearBedCache}
\alias{clearBedCache}
\title{Clear the BED cache SQLite database}
\usage{
clearBedCache(queries = NULL, force = FALSE, hard = FALSE, verbose = FALSE)
}
\arguments{
\item{queries}{a character vector of the names of queries to remove.
If NULL all queries are removed.}

\item{force}{if TRUE clear the BED cache table even if cache
file is not found}

\item{hard}{if TRUE remove everything in cache without checking file names}

\item{verbose}{display some information during the process}
}
\description{
Clear the BED cache SQLite database
}
\seealso{
\link{lsBedCache}
}
