% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareBedInstances.R
\name{compareBedInstances}
\alias{compareBedInstances}
\title{Compare 2 BED database instances}
\usage{
compareBedInstances(connections)
}
\arguments{
\item{connections}{a numeric vector of length 1 or 2 providing connections
from \link{lsBedConnections} to be compared.}
}
\value{
If only one connection is provided, the function returns a list
with information about BEID and platforms available for the connection
along with DB version information.
If two connections are provided the same information as above is provided
for the 2 connection named V1 and V2 in that order. In addition,
differences observed between the 2 instances are reported for BEID and
platforms.
}
\description{
Compare 2 BED database instances
}
\details{
The current connection is restored when exiting this function.
}
