% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convBeIdLists.R
\name{convBeIdLists}
\alias{convBeIdLists}
\title{Converts lists of BE IDs}
\usage{
convBeIdLists(idList, ...)
}
\arguments{
\item{idList}{a list of IDs lists}

\item{...}{params for the \code{\link{convBeIds}} function}
}
\value{
a list of \code{\link{convBeIds}} ouput ids
}
\description{
Converts lists of BE IDs
}
\examples{
\dontrun{
convBeIdLists(
   idList=list(a=c("10", "100"), b=c("1000")),
   from="Gene",
   from.source="EntrezGene",
   from.org="human",
   to.source="Ens_gene"
)
}

}
\seealso{
\code{\link{convBeIds}}, \code{\link{convDfBeIds}}
}
