% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convDfBeIds.R
\name{convDfBeIds}
\alias{convDfBeIds}
\title{Add BE ID conversion to a data frame}
\usage{
convDfBeIds(df, idCol = NULL, ...)
}
\arguments{
\item{df}{the data.frame to be converted}

\item{idCol}{the column in which ID to convert are. If NULL (default)
the row names are taken.}

\item{...}{params for the \code{\link{convBeIds}} function}
}
\value{
a data.frame with converted IDs
}
\description{
Add BE ID conversion to a data frame
}
\examples{
\dontrun{
toConv <- data.frame(a=1:2, b=3:4)
rownames(toConv) <- c("10", "100")
convDfBeIds(
   df=toConv,
   from="Gene",
   from.source="EntrezGene",
   from.org="human",
   to.source="Ens_gene"
)
}

}
\seealso{
\code{\link{convBeIds}}, \code{\link{convBeIdLists}}
}
