% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpNcbiDb.R
\name{dumpNcbiDb}
\alias{dumpNcbiDb}
\title{Feeding BED: Dump tables from the NCBI gene DATA}
\usage{
dumpNcbiDb(taxOfInt, reDumpThr, toLoad = c("gene_info", "gene2ensembl",
  "gene2unigene", "gene2vega", "gene_group", "gene_orthologs",
  "gene_history", "gene2refseq"), env = parent.frame(n = 1), curDate)
}
\arguments{
\item{taxOfInt}{the organism to download (e.g. "9606").}

\item{reDumpThr}{time difference threshold between 2 downloads}

\item{toLoad}{the list of tables to load}

\item{env}{the R environment in which to load the tables when downloaded}

\item{curDate}{current date as given by \code{\link{Sys.Date}}}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
