% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreConvPath.R
\name{exploreConvPath}
\alias{exploreConvPath}
\title{Explore the shortest convertion path between two identifiers}
\usage{
exploreConvPath(from.id, from, from.source, to.id, to = from,
  to.source = from.source, verbose = FALSE)
}
\arguments{
\item{from.id}{the first identifier}

\item{from}{the type of entity: \code{listBe()} or Probe}

\item{from.source}{the identifier source: database or platform}

\item{to.id}{the first identifier}

\item{to}{the type of entity: \code{listBe()} or Probe}

\item{to.source}{the identifier source: database or platform}

\item{verbose}{if TRUE the cypher query is shown}
}
\description{
This function uses visNetwork to draw the shortest convertion path between
two identifiers (including ProbeID)
}
\examples{
\dontrun{
exploreConvPath(
   from.id="ENST00000413465", from="Transcript", from.source="Ens_transcript",
   to.id="ENSMUST00000108658", to="Transcript", to.source="Ens_transcript"
)
}

}
