% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filterByBEID}
\alias{filterByBEID}
\title{Filter an object to keep only a set of BEIDs}
\usage{
filterByBEID(x, toKeep, ...)
}
\arguments{
\item{x}{an object representing a collection of BEID (e.g. BEIDList)}

\item{toKeep}{a vector of elements to keep}

\item{...}{method specific parameters}
}
\description{
Filter an object to keep only a set of BEIDs
}
