% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBeids.R
\name{findBeids}
\alias{findBeids}
\title{Find Biological Entity identifiers}
\usage{
findBeids(toGene = TRUE, ...)
}
\arguments{
\item{toGene}{focus on gene entities (default=TRUE): matches from other
BE are converted to genes.}

\item{...}{parameters for \link{beidsServer}}
}
\value{
NULL if not any result, or a data.frame with the selected
values and the following column:
\itemize{
\item \strong{value}: the BE identifier
\item \strong{preferred}: preferred identifier for the same BE in the same scope
\item \strong{be}: the type of biological entity
\item \strong{source}: the source of the identifier
\item \strong{organism}: the organism of the BE
\item \strong{canonical} (if toGene==TRUE): canonical gene product? (if known)
\item \strong{symbol}: the symbol of the identifier (if any)
}
}
\description{
Find Biological Entity identifiers
}
