% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{focusOnScope}
\alias{focusOnScope}
\title{Focus a BE related object on a specific identifier (BEID) scope}
\usage{
focusOnScope(x, ...)
}
\arguments{
\item{x}{the object to be focused}

\item{...}{additional parameters to the BEID conversion function}

\item{be}{the type of biological entity to focus on}

\item{source}{the source of BEID to focus on}

\item{organism}{the organism of BEID to focus on}

\item{restricted}{if TRUE (default) the BEID are limited to current version
of the source}

\item{prefFilter}{if TRUE (default) the BEID are limited to prefered
identifiers when they exist}
}
\value{
Depends on the class of x
}
\description{
Focus a BE related object on a specific identifier (BEID) scope
}
