% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{forgetBedConnection}
\alias{forgetBedConnection}
\title{Forget a BED connection}
\usage{
forgetBedConnection(connection)
}
\arguments{
\item{connection}{the id of the connection to forget.}
}
\description{
Forget a BED connection
}
\seealso{
\code{\link{lsBedConnections}},
\code{\link{checkBedConn}}, \code{\link{connectToBed}}
}
