% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBePath.R
\name{genBePath}
\alias{genBePath}
\title{Construct CQL sub-query to map 2 biological entity}
\usage{
genBePath(from, to)
}
\arguments{
\item{from}{one biological entity (BE)}

\item{to}{one biological entity (BE)}
}
\value{
A character value corresponding to the sub-query
}
\description{
Internal use
}
\seealso{
\code{\link{genProbePath}}, \code{\link{listBe}}
}
