% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_to_all_scopes.R
\name{geneIDs_to_all_scopes}
\alias{geneIDs_to_all_scopes}
\title{Find all BEID and ProbeID}
\usage{
geneIDs_to_all_scopes(
  geneids,
  source,
  organism,
  entities = NULL,
  canonical_symbols = TRUE
)
}
\description{
Find all BEID and ProbeID
}
