% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genProbePath.R
\name{genProbePath}
\alias{genProbePath}
\title{Identify the biological entity (BE) targeted by probes and
construct the CQL sub-query to map probes to the BE}
\usage{
genProbePath(platform)
}
\arguments{
\item{platform}{the platform of the probes}
}
\value{
A character value corresponding to the sub-query.
The \code{attr(,"be")} correspond to the BE targeted by probes
}
\description{
Internal use
}
\seealso{
\link{genBePath}, \link{listPlatforms}
}
