% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllBeIdSources.R
\name{getAllBeIdSources}
\alias{getAllBeIdSources}
\title{List all the source databases of BE identifiers whatever the BE type}
\usage{
getAllBeIdSources(recache = FALSE)
}
\arguments{
\item{recache}{boolean indicating if the CQL query should be run even if
the table is already in cache}
}
\value{
A data.frame indicating the BE related to the ID source (database).
}
\description{
List all the source databases of BE identifiers whatever the BE type
}
\seealso{
\link{listBeIdSources}, \link{listPlatforms}
}
