% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBeIdConvTable.R
\name{getBeIdConvTable}
\alias{getBeIdConvTable}
\title{Get a conversion table between biological entity (BE) identifiers}
\usage{
getBeIdConvTable(from, to = from, from.source, to.source, organism,
  restricted = TRUE, entity = TRUE, verbose = FALSE,
  recache = FALSE, filter = NULL)
}
\arguments{
\item{from}{one BE or "Probe"}

\item{to}{one BE or "Probe"}

\item{from.source}{the from BE ID database if BE or
the from probe platform if Probe}

\item{to.source}{the to BE ID database if BE or
the to probe platform if Probe}

\item{organism}{organism name}

\item{restricted}{boolean indicating if the results should be restricted to
current version of to BEID db. If FALSE former BEID are also returned:
\strong{Depending on history it can take a very long time to return
a very large result!}}

\item{entity}{boolean indicating if the technical ID of to BE should be
returned}

\item{verbose}{boolean indicating if the CQL query should be displayed}

\item{recache}{boolean indicating if the CQL query should be run even if
the table is already in cache}

\item{filter}{character vector on which to filter from IDs.
If NULL (default), the result is not filtered:
all from IDs are taken into account. Filtering is case insensitive.}
}
\value{
a data.frame mapping BE IDs with the
following fields:
\describe{
 \item{from}{the from BE ID}
 \item{to}{the to BE ID}
 \item{entity}{(optional) the technical ID of to BE}
}
}
\description{
Get a conversion table between biological entity (BE) identifiers
}
\examples{
\dontrun{
getBeIdConvTable(
    from="Gene", from.source="EntrezGene",
    to.source="Ens_gene",
    organism="human"
)
}

}
\seealso{
\code{\link{getHomTable}}, \code{\link{listBe}},
\code{\link{listPlatforms}}, \code{\link{listBeIdSources}}
}
