% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBeIdDescription.R
\name{getBeIdDescription}
\alias{getBeIdDescription}
\title{Get description of Biological Entity identifiers}
\usage{
getBeIdDescription(ids, be, source, organism, ...)
}
\arguments{
\item{ids}{list of identifiers}

\item{be}{one BE}

\item{source}{the BE ID database}

\item{organism}{organism name}

\item{...}{further arguments for
\code{\link{getBeIdNames}} and \code{\link{getBeIdSymbols}} functions}
}
\value{
a data.frame providing for each BE IDs
(row.names are provided BE IDs):
\describe{
\item{id}{the BE ID}
\item{symbol}{the BE symbol}
\item{name}{the corresponding name}
}
}
\description{
This description can be used for annotating tables or graph based on BE IDs.
}
\examples{
\dontrun{
getBeIdDescription(
   ids=c("10", "100"),
   be="Gene",
   source="EntrezGene",
   organism="human"
)
}

}
\seealso{
\code{\link{getBeIdNames}},
\code{\link{getBeIdSymbols}}
}
