% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBeIdNames.R
\name{getBeIdNames}
\alias{getBeIdNames}
\title{Get names of Biological Entity identifiers}
\usage{
getBeIdNames(ids, limForCache = 4000, ...)
}
\arguments{
\item{ids}{list of identifiers}

\item{limForCache}{if there are more ids than limForCache results are
collected for all IDs (beyond provided ids) and cached for futur queries.
If not, results are collected only for provided ids and not cached.}

\item{...}{params for the \code{\link{getBeIdNameTable}} function}
}
\value{
a data.frame mapping BE IDs and names with the
following fields:
\describe{
 \item{id}{the BE ID}
 \item{name}{the corresponding name}
 \item{direct}{true if the name is directly related to the BE ID}
 \item{entity}{(optional) the technical ID of to BE}
}
}
\description{
Get names of Biological Entity identifiers
}
\examples{
\dontrun{
getBeIdNames(
   ids=c("10", "100"),
   be="Gene",
   source="EntrezGene",
   organism="human"
)
}

}
\seealso{
\code{\link{getBeIdNameTable}},
\code{\link{getBeIdSymbols}}
}
