% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBeIdNameTable.R
\name{getBeIdNameTable}
\alias{getBeIdNameTable}
\title{Get a table of biological entity (BE) identifiers and names}
\usage{
getBeIdNameTable(
  be,
  source,
  organism,
  restricted,
  entity = TRUE,
  verbose = FALSE,
  recache = FALSE,
  filter = NULL
)
}
\arguments{
\item{be}{one BE}

\item{source}{the BE ID database}

\item{organism}{organism name}

\item{restricted}{boolean indicating if the results should be restricted to
direct names}

\item{entity}{boolean indicating if the technical ID of BE should be
returned}

\item{verbose}{boolean indicating if the CQL query should be displayed}

\item{recache}{boolean indicating if the CQL query should be run even if
the table is already in cache}

\item{filter}{character vector on which to filter id. If NULL (default),
the result is not filtered: all IDs are taken into account.}
}
\value{
a data.frame with the
following fields:
\itemize{
\item \strong{id}: the from BE ID
\item \strong{name}: the BE name
\item \strong{direct}: false if the symbol is not directly associated to the BE ID
\item \strong{preferred}: true if the id is the preferred identifier for the BE
\item \strong{entity}: (optional) the technical ID of to BE
}
}
\description{
Get a table of biological entity (BE) identifiers and names
}
\examples{
\dontrun{
getBeIdNameTable(
   be="Gene",
   source="EntrezGene",
   organism="human"
)
}

}
\seealso{
\link{getBeIdNames}, \link{getBeIdSymbolTable}
}
