% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBeIds.R
\name{getBeIds}
\alias{getBeIds}
\title{Get biological entities identifiers}
\usage{
getBeIds(be = c(listBe(), "Probe"), source, organism = NA, restricted,
  entity = TRUE, attributes = NULL, verbose = FALSE,
  recache = FALSE, filter = NULL, limForCache = 4000)
}
\arguments{
\item{be}{one BE or "Probe"}

\item{source}{the BE ID database or "Symbol" if BE or
the probe platform if Probe}

\item{organism}{organism name}

\item{restricted}{boolean indicating if the results should be restricted to
current version of to BEID db. If FALSE former BEID are also returned.}

\item{entity}{boolean indicating if the technical ID of BE should be
returned}

\item{attributes}{a character vector listing attributes that should be
returned.}

\item{verbose}{boolean indicating if the CQL query should be displayed}

\item{recache}{boolean indicating if the CQL query should be run even if
the table is already in cache}

\item{filter}{character vector on which to filter id. If NULL (default),
the result is not filtered: all IDs are taken into account.}

\item{limForCache}{if there are more filter than limForCache results are
collected for all IDs (beyond provided ids) and cached for futur queries.
If not, results are collected only for provided ids and not cached.}
}
\value{
a data.frame mapping BE IDs with the
following fields:
\describe{
 \item{id}{the BE ID}
 \item{BE}{IF entity is TRUE the technical ID of BE}
 \item{db.version}{IF be is not "Probe" and source not "Symbol"
 the version of the DB}
 \item{db.deprecated}{IF be is not "Probe" and source not "Symbol"
 a value if the BE ID is deprecated or FALSE if it's not}
 \item{canonical}{IF source is "Symbol" TRUE if the symbol is canonical}
 \item{organism}{IF be is "Probe" the organism of the targeted BE}
}
If attributes are part of the query, additional columns for each of them.
Scope ("be", "source" and "organism") is provided as a named list
in the "scope" attributes: \code{attr(x, "scope")}
}
\description{
Get biological entities identifiers
}
\examples{
\dontrun{
beids <- getBeIds(be="Gene", source="EntrezGene", organism="human", restricted=TRUE)
}

}
\seealso{
\code{\link{listPlatforms}}, \code{\link{listBeIdSources}}
}
