% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBeIdSymbols.R
\name{getBeIdSymbols}
\alias{getBeIdSymbols}
\title{Get symbols of Biological Entity identifiers}
\usage{
getBeIdSymbols(ids, limForCache = 4000, ...)
}
\arguments{
\item{ids}{list of identifiers}

\item{limForCache}{if there are more ids than limForCache. Results are
collected for all IDs (beyond provided ids) and cached for futur queries.
If not, results are collected only for provided ids and not cached.}

\item{...}{params for the \code{\link{getBeIdSymbolTable}} function}
}
\value{
a data.frame with the
following fields:
\describe{
\item{id}{the from BE ID}
\item{symbol}{the BE symbol}
\item{canonical}{true if the symbol is canonical for the direct BE ID}
\item{direct}{false if the symbol is not directly associated to the BE ID}
\item{entity}{(optional) the technical ID of to BE}
}
}
\description{
Get symbols of Biological Entity identifiers
}
\examples{
\dontrun{
getBeIdSymbols(
   ids=c("10", "100"),
   be="Gene",
   source="EntrezGene",
   organism="human"
)
}

}
\seealso{
\code{\link{getBeIdSymbolTable}},
\code{\link{getBeIdNames}}
}
