% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBeIdURL.R
\name{getBeIdURL}
\alias{getBeIdURL}
\title{Get a reference URL for a BE ID}
\usage{
getBeIdURL(id, database)
}
\arguments{
\item{id}{the BE ID}

\item{database}{the database from which the ID has been taken}
}
\value{
A character vector of length one corresponding to the relevant URL.
NA is returned is there is no URL corresponding to the provided database.
}
\description{
Get a reference URL for a BE ID
}
\examples{
\dontrun{
getBeIdURL("100", "EntrezGene")
}

}
