% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHomTable.R
\name{getHomTable}
\alias{getHomTable}
\title{Get gene homologs between 2 organisms}
\usage{
getHomTable(from.org, to.org, from.source = "Ens_gene",
  to.source = from.source, verbose = FALSE, recache = FALSE,
  filter = NULL)
}
\arguments{
\item{from.org}{organism name}

\item{to.org}{organism name}

\item{from.source}{the from gene ID database}

\item{to.source}{the to gene ID database}

\item{verbose}{boolean indicating if the CQL query should be displayed}

\item{recache}{boolean indicating if the CQL query should be run even if
the table is already in cache}

\item{filter}{character vector on which to filter from IDs.
If NULL (default), the result is not filtered:
all from IDs are taken into account.}
}
\value{
a data.frame mapping gene IDs with the
following fields:
\describe{
 \item{from}{the from gene ID}
 \item{to}{the to gene ID}
}
}
\description{
Get gene homologs between 2 organisms
}
\examples{
\dontrun{
getHomTable(
   from.org="human",
   to.org="mouse"
)
}

}
\seealso{
\code{\link{getBeIdConvTable}}
}
