% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTargetedBe.R
\name{getTargetedBe}
\alias{getTargetedBe}
\title{Identify the biological entity (BE) targeted by probes}
\usage{
getTargetedBe(platform)
}
\arguments{
\item{platform}{the platform of the probes}
}
\value{
The BE targeted by the platform
}
\description{
Identify the biological entity (BE) targeted by probes
}
\examples{
\dontrun{
getTargetedBe("GPL570")
}

}
\seealso{
\link{listPlatforms}
}
