% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxId.R
\name{getTaxId}
\alias{getTaxId}
\title{Get taxonomy ID of an organism name}
\usage{
getTaxId(name)
}
\arguments{
\item{name}{the name of the organism}
}
\value{
A vector of taxonomy ID
}
\description{
Get taxonomy ID of an organism name
}
\examples{
\dontrun{
getTaxId("human")
}

}
\seealso{
\link{getOrgNames}, \link{listOrganisms}
}
