% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniprot.R
\name{getUniprot}
\alias{getUniprot}
\title{Feeding BED: Download Uniprot information in BED}
\usage{
getUniprot(organism, release)
}
\arguments{
\item{organism}{character vector of 1 element corresponding to the organism
of interest (e.g. "Homo sapiens")}

\item{release}{the release of interest (check if already downloaded)}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
