% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessIdOrigin.R
\name{guessIdOrigin}
\alias{guessIdOrigin}
\title{Guess biological entity (BE), database source and organism of a vector
of identifiers.}
\usage{
guessIdOrigin(ids, tcLim = 100)
}
\arguments{
\item{ids}{a character vector of identifiers}

\item{tcLim}{number of identifiers to check to guess origin for the whole set.
Inf ==> no limit.}
}
\value{
a list: \itemize{
 \item{be: a character vector of length 1 providing the best BE guess}
 \item{source: a character vector of length 1 providing the best source guess}
 \item{organism: a character vector of length 1 providing the best organism guess}
}
"details" attribute (\code{attr(x, "details")}) is a data frame providing numbers
supporting the guess
}
\description{
Guess biological entity (BE), database source and organism of a vector
of identifiers.
}
\examples{
\dontrun{
guessIdOrigin(ids=c("10", "100"))
}

}
