% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identicalScopes.R
\name{identicalScopes}
\alias{identicalScopes}
\title{Check if two objects have the same BEID scope}
\usage{
identicalScopes(x, y)
}
\arguments{
\item{x}{the object to test}

\item{y}{the object to test}
}
\value{
A logical indicating if the 2 scopes are identical
}
\description{
Check if two objects have the same BEID scope
}
