% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEIDList.R
\name{is.BEIDList}
\alias{is.BEIDList}
\title{Check if the provided object is a \link{BEIDList}}
\usage{
is.BEIDList(x)
}
\arguments{
\item{x}{the object to check}
}
\value{
A logical value
}
\description{
Check if the provided object is a \link{BEIDList}
}
