% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBe.R
\name{listBe}
\alias{listBe}
\title{Lists all the biological entities (BE) available in the BED database}
\usage{
listBe()
}
\value{
A character vector of biological entities (BE)
}
\description{
Lists all the biological entities (BE) available in the BED database
}
\seealso{
\link{listPlatforms}, \link{listBeIdSources},
\link{listOrganisms}
}
