% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDBAttributes.R
\name{listDBAttributes}
\alias{listDBAttributes}
\title{List all attributes provided by a BEDB}
\usage{
listDBAttributes(dbname)
}
\arguments{
\item{dbname}{the name of the database}
}
\value{
A character vector of attribute names
}
\description{
List all attributes provided by a BEDB
}
