% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBedModel.R
\name{loadBedOtherIndexes}
\alias{loadBedOtherIndexes}
\title{Feeding BED: Load additional indexes in neo4j}
\usage{
loadBedOtherIndexes()
}
\description{
Not exported to avoid unintended modifications of the DB.
}
