% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheBedResult.R
\name{loadBedResult}
\alias{loadBedResult}
\title{Get a BED query result from cache}
\usage{
loadBedResult(name)
}
\arguments{
\item{name}{the name of the query}
}
\description{
Internal use
}
\seealso{
\link{cacheBedCall}, \link{cacheBedResult}
}
