% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBENames.R
\name{loadBENames}
\alias{loadBENames}
\title{Feeding BED: Load names associated to BEIDs}
\usage{
loadBENames(d, be = "Gene", dbname)
}
\arguments{
\item{d}{a data.frame with information about the names
to be loaded. It should contain the following fields: "id", "name"
and "canonical" (optional).}

\item{be}{a character corresponding to the BE type (default: "Gene")}

\item{dbname}{the DB of BEID}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
