% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBESymbols.R
\name{loadBESymbols}
\alias{loadBESymbols}
\title{Feeding BED: Load symbols associated to BEIDs}
\usage{
loadBESymbols(d, be = "Gene", dbname)
}
\arguments{
\item{d}{a data.frame with information about the symbols
to be loaded. It should contain the following fields: "id", "symbol"
and "canonical" (optional).}

\item{be}{a character corresponding to the BE type (default: "Gene")}

\item{dbname}{the DB of BEID}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
