% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBEVersion.R
\name{loadBEVersion}
\alias{loadBEVersion}
\title{Feeding BED: Load biological entities in BED with information about
DB version}
\usage{
loadBEVersion(d, be = "Gene", dbname, taxId = NA, onlyId = FALSE)
}
\arguments{
\item{d}{a data.frame with information about the entities to be loaded.
It should contain the following fields: "id", "version" and "deprecated".}

\item{be}{a character corresponding to the BE type (default: "Gene")}

\item{dbname}{the DB from which the BE ID are taken}

\item{taxId}{the taxonomy ID of the BE organism}

\item{onlyId}{a logical. If TRUE, only an BEID is created and not the
corresponding BE.}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
