% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCodesFor.R
\name{loadCodesFor}
\alias{loadCodesFor}
\title{Feeding BED: Load correspondance between genes and objects as coding events}
\usage{
loadCodesFor(d, gdb, odb)
}
\arguments{
\item{d}{a data.frame with information about the coding events.
It should contain the following fields: "gid" and "oid"}

\item{gdb}{the DB of Gene IDs}

\item{odb}{the DB of Object IDs}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
