% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCorrespondsTo.R
\name{loadCorrespondsTo}
\alias{loadCorrespondsTo}
\title{Feeding BED: Load correspondances between BE IDs}
\usage{
loadCorrespondsTo(d, db1, db2, be = "Gene")
}
\arguments{
\item{d}{a data.frame with information about the correspondances
to be loaded. It should contain the following fields: "id1" and "id2".}

\item{db1}{the DB of id1}

\item{db2}{the DB of id2}

\item{be}{a character corresponding to the BE type (default: "Gene")}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
