% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHistory.R
\name{loadHistory}
\alias{loadHistory}
\title{Feeding BED: Load history of BEIDs}
\usage{
loadHistory(d, dbname, be = "Gene")
}
\arguments{
\item{d}{a data.frame with information about the history.
It should contain the following fields: "old" and "new".}

\item{dbname}{the DB of BEID}

\item{be}{a character corresponding to the BE type (default: "Gene")}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
