% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadIsExpressedAs.R
\name{loadIsExpressedAs}
\alias{loadIsExpressedAs}
\title{Feeding BED: Load correspondance between genes and transcripts as
expression events}
\usage{
loadIsExpressedAs(d, gdb, tdb)
}
\arguments{
\item{d}{a data.frame with information about the expression events.
It should contain the following fields: "gid", "tid"
and "canonical" (optional).}

\item{gdb}{the DB of Gene IDs}

\item{tdb}{the DB of Transcript IDs}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
