% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadIsTranslatedIn.R
\name{loadIsTranslatedIn}
\alias{loadIsTranslatedIn}
\title{Feeding BED: Load correspondance between transcripts and peptides as
translation events}
\usage{
loadIsTranslatedIn(d, tdb, pdb)
}
\arguments{
\item{d}{a data.frame with information about the translation events.
It should contain the following fields: "tid", "pid"
and "canonical" (optional).}

\item{tdb}{the DB of Transcript IDs}

\item{pdb}{the DB of Peptide IDs}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
