% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLuceneIndexes.R
\name{loadLuceneIndexes}
\alias{loadLuceneIndexes}
\title{Feeding BED: Create Lucene indexes in neo4j}
\usage{
loadLuceneIndexes()
}
\description{
Not exported to avoid unintended modifications of the DB.
}
