% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMBObjects.R
\name{loadMBObjects}
\alias{loadMBObjects}
\title{Feeding BED: Load in BED objects from Clarivate Analytics
(formerly Thomson-Reuters) MetaBase}
\usage{
loadMBObjects(orgOfInt)
}
\arguments{
\item{orgOfInt}{organims to be taken into account
(e.g. c("Homo sapiens", "Mus musculus", "Rattus norvegicus"))}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
