% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNCBIEntrezGOFunctions.R
\name{loadNCBIEntrezGOFunctions}
\alias{loadNCBIEntrezGOFunctions}
\title{Feeding BED: Load in BED GO functions associated to Entrez gene IDs
from NCBI}
\usage{
loadNCBIEntrezGOFunctions(organism, reDumpThr = 1e+05, ddir, curDate)
}
\arguments{
\item{organism}{character vector of 1 element corresponding to the organism
of interest (e.g. "Homo sapiens")}

\item{reDumpThr}{time difference threshold between 2 downloads}

\item{ddir}{path to the directory where the data should be saved}

\item{curDate}{current date as given by \link{Sys.Date}}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
