% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOrganisms.R
\name{loadOrganisms}
\alias{loadOrganisms}
\title{Feeding BED: Load organisms in BED}
\usage{
loadOrganisms(d)
}
\arguments{
\item{d}{a data.frame with 2 columns named "tax_id" and "name_txt" providing
the taxonomic ID for each organism name}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
