% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPlf.R
\name{loadPlf}
\alias{loadPlf}
\title{Feeding BED: Load a probes platform}
\usage{
loadPlf(name, description, be)
}
\arguments{
\item{name}{the name of the platform}

\item{description}{a description of the platform}

\item{be}{the type of BE targeted by the platform}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
