% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProbes.R
\name{loadProbes}
\alias{loadProbes}
\title{Feeding BED: Load probes targeting BE IDs}
\usage{
loadProbes(d, be = "Transcript", platform, dbname)
}
\arguments{
\item{d}{a data.frame with information about the entities to be loaded.
It should contain the following fields: "id" and "probeID".}

\item{be}{a character corresponding to the BE
targeted by the probes (default: "Transcript")}

\item{platform}{the plateform gathering the probes}

\item{dbname}{the DB from which the BE ID are taken}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
