% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{metadata<-}
\alias{metadata<-}
\title{Set object metadata}
\usage{
metadata(x) <- value
}
\arguments{
\item{x}{an object representing a collection of BEID (e.g. BEIDList)}

\item{value}{a data.frame with rownames or
a column "\strong{.lname}" all in names of l.}
}
\description{
Set object metadata
}
