% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{metadata}
\alias{metadata}
\title{Get object metadata}
\usage{
metadata(x, ...)
}
\arguments{
\item{x}{an object representing a collection of BEID (e.g. BEIDList)}

\item{...}{method specific parameters}
}
\description{
Get object metadata
}
