% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerBEDB.R
\name{registerBEDB}
\alias{registerBEDB}
\title{Feeding BED: Register a database of biological entities in BED DB}
\usage{
registerBEDB(name, description = NA, currentVersion = NA, idURL = NA)
}
\arguments{
\item{name}{of the database (e.g. "Ens_gene")}

\item{description}{a short description of the database (e.g. "Ensembl gene")}

\item{currentVersion}{the version taken into account in BED (e.g. 83)}

\item{idURL}{the URL template to use to retrieve id information. A '\%s'
corresponding to the ID should be present in this character vector of
length one.}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
