% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{scope}
\alias{scope}
\title{Get the BEID scope of an object}
\usage{
scope(x, ...)
}
\arguments{
\item{x}{an object representing a collection of BEID (e.g. BEIDList)}

\item{...}{method specific parameters}
}
\description{
Get the BEID scope of an object
}
