% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{scopes}
\alias{scopes}
\title{Get the BEID scopes of an object}
\usage{
scopes(x, ...)
}
\value{
A tibble with 4 columns:
\itemize{
\item be
\item source
\item organism
\item Freq
}
}
\description{
Get the BEID scopes of an object
}
