% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_beid.R
\name{search_beid}
\alias{search_beid}
\title{Search a BEID}
\usage{
search_beid(x)
}
\arguments{
\item{x}{a character value to search}
}
\value{
NULL if there is not any match or
a tibble with the following columns:
\itemize{
\item \strong{Value}: the matching term
\item \strong{From}: the type of the matched term (e.g. BESymbol, GeneID...)
\item \strong{BE}: the matching biological entity (BE)
\item \strong{BEID}: the BE identifier
\item \strong{Database}: the BEID reference database
\item \strong{Preferred}: TRUE if the BEID is considered as a preferred identifier
\item \strong{Symbol}: BEID canonical symbol
\item \strong{Name}: BEID name
\item \strong{Entity}: technical BE identifier
\item \strong{GeneID}: Corresponding gene identifier
\item \strong{Gene_DB}: Gene ID database
\item \strong{Preferred_gene}: TRUE if the GeneID is considered as a preferred identifier
\item \strong{Gene_symbol}: Gene symbol
\item \strong{Gene_name}: Gene name
\item \strong{Gene_entity}: technical gene identifier
\item \strong{Organism}: gene organism (scientific name)
}
}
\description{
Search a BEID
}
