% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBedVersion.R
\name{setBedVersion}
\alias{setBedVersion}
\title{Feeding BED: Set the BED version}
\usage{
setBedVersion(bedInstance, bedVersion)
}
\arguments{
\item{bedInstance}{instance of BED to be set}

\item{bedVersion}{version of BED to be set}
}
\description{
Not exported to avoid unintended modifications of the DB.
This function is used when modifying the BED content.
}
