% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{BEIDs}
\alias{BEIDs}
\title{Get the BEIDs from an object}
\usage{
BEIDs(x, ...)
}
\arguments{
\item{x}{an object representing a collection of BEID (e.g. BEIDList)}

\item{...}{method specific parameters}
}
\value{
A tibble with at least 4 columns:
\itemize{
\item value
\item be
\item source
\item organism
\item ...
}
}
\description{
Get the BEIDs from an object
}
