% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpNcbiTax.R
\name{dumpNcbiTax}
\alias{dumpNcbiTax}
\title{Feeding BED: Dump tables with taxonomic information from NCBI}
\usage{
dumpNcbiTax(
  reDumpThr,
  ddir,
  toDump = c("names.dmp"),
  env = parent.frame(n = 1),
  curDate
)
}
\arguments{
\item{reDumpThr}{time difference threshold between 2 downloads}

\item{ddir}{path to the directory where the data should be saved}

\item{toDump}{the list of tables to load}

\item{env}{the R environment in which to load the tables when downloaded}

\item{curDate}{current date as given by \link{Sys.Date}}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
