% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpUniprotDb.R
\name{dumpUniprotDb}
\alias{dumpUniprotDb}
\title{Feeding BED: Dump and preprocess flat data files from Uniprot}
\usage{
dumpUniprotDb(
  taxOfInt,
  divOfInt,
  release,
  ddir,
  ftp = "ftp://ftp.expasy.org/databases/uniprot",
  env = parent.frame(n = 1)
)
}
\arguments{
\item{taxOfInt}{the organism of interest (e.g., "9606" for human,
"10090" for mouse or "10116" for rat)}

\item{divOfInt}{the taxonomic division to which the organism belong
(e.g., "human", "rodents", "mammals", "vertebrates")}

\item{release}{the release of interest (check if already downloaded)}

\item{ddir}{path to the directory where the data should be saved}

\item{ftp}{location of the ftp site}

\item{env}{the R environment in which to load the tables when built}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
