% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpUniprotDb.R
\name{dumpUniprotDb}
\alias{dumpUniprotDb}
\title{Feeding BED: Dump and preprocess flat dat files fro Uniprot}
\usage{
dumpUniprotDb(
  taxOfInt,
  release,
  ddir,
  ftp = "ftp://ftp.expasy.org/databases/uniprot",
  env = parent.frame(n = 1)
)
}
\arguments{
\item{taxOfInt}{the organism of interest. Only human ("9606"),
mouse ("10090") and rat ("10116") are supported}

\item{release}{the release of interest (check if already downloaded)}

\item{ddir}{path to the directory where the data should be saved}

\item{ftp}{location of the ftp site}

\item{env}{the R environment in which to load the tables when built}
}
\description{
Not exported to avoid unintended modifications of the DB.
}
