% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{focusOnScope}
\alias{focusOnScope}
\title{Focus a BE related object on a specific identifier (BEID) scope}
\usage{
focusOnScope(x, ...)
}
\arguments{
\item{x}{an object representing a collection of BEID (e.g. BEIDList)}

\item{...}{method specific parameters for BEID conversion}
}
\value{
Depends on the class of x
}
\description{
Focus a BE related object on a specific identifier (BEID) scope
}
