% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{forgetBedConnection}
\alias{forgetBedConnection}
\title{Forget a BED connection}
\usage{
forgetBedConnection(connection, save = FALSE)
}
\arguments{
\item{connection}{the id of the connection to forget.}

\item{save}{a logical. Should be set to TRUE to save the updated list of
connections in the file space (default to FALSE to comply with CRAN
policies).}
}
\description{
Forget a BED connection
}
\seealso{
\link{lsBedConnections}, \link{checkBedConn}, \link{connectToBed}
}
