% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneIDs_to_all_scopes.R
\name{geneIDs_to_all_scopes}
\alias{geneIDs_to_all_scopes}
\title{Find all BEID and ProbeID}
\usage{
geneIDs_to_all_scopes(
  geneids,
  source,
  organism,
  entities = NULL,
  canonical_symbols = TRUE,
  entity_warning = TRUE
)
}
\arguments{
\item{geneids}{a character vector of gene identifiers}

\item{source}{the source of gene identifiers}

\item{organism}{the gene organism}

\item{entities}{a numeric vector of gene entity. If NULL (default),
geneids, source and organism arguments are used to identify genes.
Be carefull when using entities as these identifiers are not stable.}

\item{canonical_symbols}{return only canonical symbols (default: TRUE).}

\item{entity_warning}{by default (TRUE) a warning is shown when
the entities argument is used. Set this argument to FALSE to avoid this
warning.}
}
\value{
A tibble with the following fields:
\itemize{
\item \strong{value}: the identifier
\item \strong{be}: the type of BE
\item \strong{organism}: the BE organism
\item \strong{source}: the source of the identifier
}
}
\description{
Find all BEID and ProbeID
}
