% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOrganisms.R
\name{listOrganisms}
\alias{listOrganisms}
\title{Lists all the organisms available in the BED database}
\usage{
listOrganisms()
}
\value{
A character vector of organism scientific names
}
\description{
Lists all the organisms available in the BED database
}
\seealso{
\link{listPlatforms}, \link{listBeIdSources},
\link{listBe}, \link{getTaxId}, \link{getOrgNames}
}
