% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPlatforms.R
\name{listPlatforms}
\alias{listPlatforms}
\title{Lists all the probe platforms available in the BED database}
\usage{
listPlatforms(be = c(NA, listBe()))
}
\arguments{
\item{be}{a character vector of BE on which to focus.
if NA (default) all the BE are considered.}
}
\value{
A data.frame mapping platforms to BE with the following fields:
\itemize{
\item \strong{name}: the platform nam
\item \strong{description}: platform description
\item \strong{focus}: Targeted BE
}
}
\description{
Lists all the probe platforms available in the BED database
}
\examples{
\dontrun{
listPlatforms(be="Gene")
listPlatforms()
}

}
\seealso{
\link{listBe}, \link{listBeIdSources},
\link{listOrganisms}, \link{getTargetedBe}
}
