% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{lsBedConnections}
\alias{lsBedConnections}
\title{List all registered BED connection}
\usage{
lsBedConnections()
}
\description{
List all registered BED connection
}
\seealso{
\link{connectToBed}, \link{forgetBedConnection}, \link{checkBedConn}
}
